[FLINK-32435] Moves NoOpLeaderElection out of DocumentingDispatcherRestEndpoint
