+++ flink-runtime/src/test/java/org/apache/flink/runtime/leaderelection/NoOpLeaderElection.java
@@ -1,0 +1,40 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.flink.runtime.leaderelection;
+
+import java.util.UUID;
+
+/** {@code NoOpLeaderElection} is a no-operations implementation of {@link LeaderElection}. */
+public enum NoOpLeaderElection implements LeaderElection {
+    INSTANCE;
+
+    @Override
+    public void startLeaderElection(LeaderContender contender) {}
+
+    @Override
+    public void confirmLeadership(UUID leaderSessionID, String leaderAddress) {}
+
+    @Override
+    public boolean hasLeadership(UUID leaderSessionId) {
+        return false;
+    }
+
+    @Override
+    public void close() {}
+}
